Fixed vertical line handling
Creation of vertical lines was failing when the coordinate values were
almost the same.  The small rounding error would cause the test to fail
and the code would attempt to make an arc rather than a line.  The
underlying API call to create the arc failed.

Added “close” comparison test for vertical lines that removes the
potential for the rounding error.